Struct local_runtime::Permill

pub struct Permill(_);

A fixed point representation of a number in the range [0, 1].

Parts per Million

Implementations

impl Permill

pub const fn from_perthousand(x: u32) -> Permill

Converts a percent into Self. Equal to x / 1000.

This can be created at compile time.

impl Permill

pub const fn from_parts(parts: u32) -> Permill

From an explicitly defined number of parts per maximum of the type.

pub const fn from_percent(x: u32) -> Permill

Converts a percent into Self. Equal to x / 100.

This can be created at compile time.

pub const fn one() -> Permill

See [PerThing::one]

pub fn is_one(&self) -> bool

See [PerThing::is_one].

pub const fn zero() -> Permill

See [PerThing::zero].

pub fn is_zero(&self) -> bool

See [PerThing::is_zero].

pub const fn deconstruct(self) -> u32

See [PerThing::deconstruct].

pub fn square(self) -> Permill

See [PerThing::square].

pub fn from_float(x: f64) -> Permill

See [PerThing::from_float].

pub fn from_rational_approximation<N>(p: N, q: N) -> Permill where
N: Clone + Ord + TryInto<u32> + TryInto<u64> + Div<N, Output = N> + Rem<N, Output = N> + Add<N, Output = N> + Unsigned,
u32: Into<N>,

👎 Deprecated:
Use PerThing::from_rational instead

See [PerThing::from_rational].

pub fn from_rational<N>(p: N, q: N) -> Permill where
N: Clone + Ord + TryInto<u32> + TryInto<u64> + Div<N, Output = N> + Rem<N, Output = N> + Add<N, Output = N> + Unsigned,
u32: Into<N>,

See [PerThing::from_rational].

pub fn mul_floor<N>(self, b: N) -> N where
N: Clone + UniqueSaturatedInto<u32> + Rem<N, Output = N> + Div<N, Output = N> + Mul<N, Output = N> + Add<N, Output = N> + Unsigned,
u32: Into<N>,

See [PerThing::mul_floor].

pub fn mul_ceil<N>(self, b: N) -> N where
N: Clone + UniqueSaturatedInto<u32> + Rem<N, Output = N> + Div<N, Output = N> + Mul<N, Output = N> + Add<N, Output = N> + Unsigned,
u32: Into<N>,

See [PerThing::mul_ceil].

pub fn saturating_reciprocal_mul<N>(self, b: N) -> N where
N: Clone + UniqueSaturatedInto<u32> + Rem<N, Output = N> + Div<N, Output = N> + Mul<N, Output = N> + Add<N, Output = N> + Saturating + Unsigned,
u32: Into<N>,

See [PerThing::saturating_reciprocal_mul].

pub fn saturating_reciprocal_mul_floor<N>(self, b: N) -> N where
N: Clone + UniqueSaturatedInto<u32> + Rem<N, Output = N> + Div<N, Output = N> + Mul<N, Output = N> + Add<N, Output = N> + Saturating + Unsigned,
u32: Into<N>,

See [PerThing::saturating_reciprocal_mul_floor].

pub fn saturating_reciprocal_mul_ceil<N>(self, b: N) -> N where
N: Clone + UniqueSaturatedInto<u32> + Rem<N, Output = N> + Div<N, Output = N> + Mul<N, Output = N> + Add<N, Output = N> + Saturating + Unsigned,
u32: Into<N>,

See [PerThing::saturating_reciprocal_mul_ceil].