Crate local_runtime

Modules

api

genesis_config

Structs

AccountMappingStorageFee

AdjustmentFactorGetter

Simple wrapper for fetching current native transaction fee weight fee multiplier.

AdjustmentVariable

AssetAccountDeposit

AssetDeposit

AssetsStringLimit

BaseNativeCurrencyPrice

BlockGasLimit

EVM gas limit

BlockHashCount

ChainId

Ethereum-compatible chain_id:

CodeHashLockupDepositPercent

CommunityCouncilCallFilter

CommunityCouncilMaxMembers

CommunityTreasuryAccountId

CommunityTreasuryPalletId

CouncilMaxMembers

DappsStakingPalletId

DefaultBaseFeePerGas

DefaultDepositLimit

DepositPerByte

DepositPerItem

ExistentialDeposit

FindAuthorTruncated

GasLimitPovSizeRatio

The amount of gas per PoV size. Value is calculated as:

InflationCycleConfig

InflationParameters

Inflation parameters.

InflationPayoutPerBlock

MainTreasuryAccount

MaxBaseFeePerGas

MaxDelegateDependencies

MaxLocks

MaxProposalWeight

MaximumMultiplier

MaximumSchedulerWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MinBaseFeePerGas

MinVestedTransfer

MinimumMultiplier

MinimumPeriod

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PostBlockAndTxnHashes

PrecompilesValue

PreimageBaseDeposit

PreimageByteDeposit

PreimageHoldReason

ProposalBond

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

We allow for 1 seconds of compute with a 2 second average block time.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SafeModeWhitelistedCalls

Calls that can bypass the safe-mode pallet.

Schedule

SessionKeys

StepLimitRatio

TargetBlockFullness

TechnicalCommitteeMaxMembers

TransactionLengthFeeFactor

TreasuryPalletId

UnvestedFundsAllowedWithdrawReasons

Version

WeightFeeFactor

WeightPerGas

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

WhitelistedCalls

Filter that only allows whitelisted runtime call to pass through dispatch precompile

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

TierThreshold

Description of tier entry requirement.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

ASSET_PRECOMPILE_ADDRESS_PREFIX

The asset precompile address prefix. Addresses that match against this prefix will be routed to Erc20AssetsPrecompileSet

AST

DAYS

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU). Given the 500ms Weight, from which 75% only are used for transactions, the total EVM execution gas limit is: GAS_PER_SECOND * 0.500 * 0.75 ~= 15_000_000.

HOURS

MICROAST

Constant values used within the runtime.

MILLIAST

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MINUTES

RUNTIME_API_VERSIONS

SLOT_DURATION

STORAGE_BYTE_FEE

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Traits

BuildStorage

Complex storage builder stuff.

Functions

deposit

Charge fee for stored bytes and items.

native_version

The version information used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Assets

AssetsConfig

Aura

AuraConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

CollectiveProxy

CommunityCouncil

CommunityCouncilConfig

CommunityCouncilMembership

CommunityCouncilMembershipConfig

CommunityTreasury

CommunityTreasuryConfig

Contracts

Council

CouncilConfig

CouncilMembership

CouncilMembershipConfig

DappStaking

DappStakingConfig

Democracy

DemocracyConfig

DynamicEvmBaseFee

EVM

EVMConfig

Ethereum

EthereumChecked

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

GrandpaId

Identity of a Grandpa authority.

Inflation

InflationConfig

LocalChainExtensions

LocalPrecompiles

Migrations

Precompiles

Preimage

Proxy

RandomnessCollectiveFlip

SafeMode

SafeModeConfig

Scheduler

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

StaticPriceProvider

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeMembership

TechnicalCommitteeMembershipConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Treasury

TreasuryConfig

TxPause

TxPauseConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UnifiedAccounts

Utility

Vesting

VestingConfig