Crate local_runtime

Modules

api

Structs

BaseFeeThreshold

BeneficiaryPayout

BlockGasLimit

EVM gas limit

BlockHashCount

BlockPerEra

CallFee

CallMagicNumber

ChainId

Ethereum-compatible chain_id:

DappsStakingPalletId

DappsStakingTvlProvider

DefaultBaseFeePerGas

DeletionQueueDepth

DeletionWeightLimit

DepositPerByte

DepositPerItem

EcdsaUnsignedPriority

ExistentialDeposit

FindAuthorTruncated

FixedGasPrice

GenesisConfig

IsActive

LocalGasWeightMapping

LocalNetworkPrecompiles

The PrecompileSet installed in the Local runtime.

MaxAuthorities

MaxEraStakeValues

MaxLocks

MaxNumberOfStakersPerContract

MaxUnlockingChunks

MaxValueSize

MinVestedTransfer

MinimumPeriod

MinimumRemainingAmount

MinimumStakingAmount

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PrecompilesValue

RegisterDeposit

RewardAmount

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

We allow for 1 seconds of compute with a 2 second average block time.

SS58Prefix

Schedule

SessionKeys

TransactionByteFee

TreasuryPalletId

UnbondingPeriod

Version

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

Event

OriginCaller

SmartContract

Multi-VM pointer to smart contract instance.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

AST

DAYS

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU). Given the 500ms Weight, from which 75% only are used for transactions, the total EVM execution gas limit is: GAS_PER_SECOND * 0.500 * 0.75 ~= 15_000_000.

HOURS

MILLIAST

Constant values used within the runtime.

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MINUTES

SLOT_DURATION

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Traits

BuildStorage

Complex storage builder stuff.

Functions

deposit

Charge fee for stored bytes and items.

native_version

The version information used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Aura

AuraConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Balance

Balance of an account.

Balances

BalancesConfig

BaseFee

BaseFeeConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

BlockReward

BlockRewardConfig

CheckedExtrinsic

Extrinsic type that has already been checked.

Contracts

DappsStaking

EVM

EVMConfig

EthCall

Ethereum

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

GrandpaId

Identity of a Grandpa authority.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

Precompiles

RandomnessCollectiveFlip

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig